+++ src/exporters/index.ts
@@ -2,3 +2,4 @@ export * from './abc.js';
 export * from './lilypond.js';
 export * from './musicxml.js';
 export * from './midi.js';
+export * from './reaper.js';